fetchai/agent/pyproject.toml
[tool.poetry]
name = "top-5-priorities-email"
version = "0.0.1"
description = "Retrieves the top 5 priorities emails from the user's inbox"
authors = ["Beniamino & Pradhumn"]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
uagents = "0.11.1"
requests = "^2.31.0"
langchain = "^0.1.12"
langchain-community = "^0.0.28"
uagents-ai-engine = "^0.1.2"
google-generativeai = "*"
fastapi = "*"
pytz = "*"